requirements.txt
Django==1.10
django-webpack-loader==0.5.0
djangorestframework==3.7.7
requests==2.10.0
six==1.11.0
urllib3==1.7.1
wheel==0.30.0
zappa==0.45.1
zappa-django-utils==0.3.1
psycopg2==2.7.4
boto==2.48.0
django-storages==1.6.5
django-cors-headers==2.1.0
jsonfield==2.0.2